test: extract integer-field validation tests
