+++ packages/number-field/test/helpers.js
@@ -1,0 +1,12 @@
+import { fire } from '@vaadin/testing-helpers';
+
+/**
+ * Emulates the user filling in something in the combo-box input.
+ *
+ * @param {Element} field
+ * @param {string} value
+ */
+export function setInputValue(field, value) {
+  field.inputElement.value = value;
+  fire(field.inputElement, 'input');
+}
